# src/asltutor/static/swagger.yaml
openapi: 3.0.0
info:
  title: ASL Tutor API
  description: Full API documentation for ASL tutor backend
  version: 1.0.0-oas3
servers:
  - url: 'http://localhost:1337/v1'
  - url: 'https://theasltutor.com/v1'
paths:
...
  /admin/stats:
    get:
      tags:
        - admin
      summary: Get top most requested words
      description: >-
        Returns a list of JSON objects with the top 'N' where 5 <= N <= 100 most
        downloaded words.
      parameters:
        - name: limit
          in: query
          description: The total number words the admin requests
          required: false
          style: form
          explode: true
          schema:
            maximum: 100
            minimum: 5
            type: integer
            format: int32
            example: 25
            default: 20
      responses:
        '200':
          description: >-
            Success: returning the dictionary information of the top 'N'
            requested words
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetDictionary'
  /admin/submissions:
    get:
      tags:
        - admin
      summary: Get a list of all submissions filtered based on certian criteria.
      description: >-
        Returns an array of all submissions. It can be filtered based on quizId
        and/or moduleId and/or userId. For example, a user can request all of
        their sumissions for any combination of moduleId or quizId. Must have at
        least one filter otherwise error
      parameters:
        - name: user
          in: query
          description: >-
            The username of the user that the admin wants all the submissions
            for
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: quiz
          in: query
          description: The quiz Id that a user wants all submissions for
          required: false
          style: form
          explode: true
          schema:
            type: string
            format: ObjectId
        - name: module
          in: query
          description: >-
            The username of the user that the admin wants all the submissions
            for
          required: false
          style: form
          explode: true
          schema:
            type: string
            format: ObjectId
        - name: submission
          in: query
          description: The Id of the submission that a user is requesting. CANNOT BE COMBINED WITH OTHER QUERIES.
          required: false
          style: form
          explode: true
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Submission(s) sucessfully returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetSubmission'
        '412':
          description: 'Failed: specify at least one filter to view submissions'
...
  '/module/quiz/id/{moduleId}':
    post:
      deprecated: true
      tags:
        - quiz
      summary: Fast create quiz
      description: >-
        If you have a complete quiz pass it here and it will populate all the
        fields for the whole quiz
      parameters:
        - name: moduleId
          in: path
          description: The module Id for the module that a user is adding a quiz to
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: 'Success: all question objects have been returned'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkCreateQuiz'
        '400':
          description: 'Failed: invalid entries'
        '415':
          description: 'Failed: Content-type must be application/json'
  '/module/quiz/id/{id}':
    post:
      tags:
        - quiz
      summary: Create quiz
      description: Create a quiz with no questions and attach it to a module
      parameters:
        - name: id
          in: path
          description: The module Id for the module that a user is adding a quiz to
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: ObjectId
      requestBody:
        description: A request containing the quiz name and details about the quiz
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostQuiz'
        required: true
      responses:
        '200':
          description: 'Success: Quiz has been created for a module'
        '400':
          description: 'Failed: invalid Id'
        '415':
          description: 'Failed: Content-type must be application/json'
        '500':
          description: 'Failed: server error'
    get:
      tags:
        - quiz
      summary: Get quiz
      description: Get a single quiz based on a given quiz Id
      parameters:
        - name: id
          in: path
          description: The quiz Id for the quiz that a user is requesting
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: 'Success: quiz object returned'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetQuiz'
        '400':
          description: 'Failed: invalid Id'
        '404':
          description: 'Failed: quiz not found'
        '500':
          description: 'Failed: server error'
  '/module/quiz/question/id/{id}':
    post:
      tags:
        - quiz
      summary: Create question
      description: Create a single question and attach it to a quiz
      parameters:
        - name: id
          in: path
          description: The quiz Id for the quiz that a user is adding a question to
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: ObjectId
      requestBody:
        description: A request containing the question text and the objectId of the word
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostQuestion'
      responses:
        '200':
          description: 'Success: all question objects have been returned'
        '400':
          description: 'Failed: invalid entries'
        '415':
          description: 'Failed: Content-type must be application/json'
        '500':
          description: 'Failed: server error'
    get:
      tags:
        - quiz
      summary: Get a question
      description: Get a single question given a question Id
      parameters:
        - name: id
          in: path
          description: The question Id for the question that a user is requesting
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: 'Success: the question object has been returned'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetQuestion'
        '400':
          description: 'Failed: invalid Id'
        '404':
          description: 'Failed: question not found'
        '500':
          description: 'Failed: server error'
  '/module/quiz/delete/id/{quizId}':
    post:
      tags:
        - quiz
      summary: Delete quiz
      description: An admin can delete an existing quiz based on a given quiz Id
      parameters:
        - name: quizId
          in: path
          description: The quiz Id for the quiz that an admin is deleting
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: 'Success: quiz has been deleted'
        '400':
          description: 'Failed: invalid id'
        '404':
          description: 'Failed: quiz not found'
  '/module/quiz/question/delete/id/{questionId}':
    post:
      tags:
        - quiz
      summary: Delete question
      description: An admin can delete an existing question based on a given question Id
      parameters:
        - name: questionId
          in: path
          description: The question Id for the question that an admin is deleting
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: 'Success: question has been deleted'
        '400':
          description: 'Failed: invalid id'
        '404':
          description: 'Failed: question not found'
  /module/quiz:
    post:
      tags:
        - submission
      summary: Submit a quiz
      description: Post a users quiz submission to the backend.
      requestBody:
        description: The submission object that the user is submiting
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostSubmission'
        required: true
      responses:
        '200':
          description: 'Success: Submission has been posted'
        '400':
          description: 'Failed: invalid Id(s)'
        '415':
          description: 'Failed: Content-type must be application/json'
components:
  schemas:
...
    PostDictionary:
      description: Word object
      type: object
      properties:
        id:
          type: string
          format: ObjectId
        word:
          type: string
        url:
          type: string
          format: uri
        in_dictionary:
          type: boolean
      example:
        in_dictionary: false
        word: test
        url: http://test.com
      required:
        - in_dictionary
        - word
    GetDictionary:
      allOf:
        - $ref: '#/components/schemas/PostDictionary'
      type: object
      properties:
        id:
          type: string
          format: ObjectId
      example:
        id: someId
    GetSubmission:
      description: Content received when a user gets a submission
      type: object
      properties:
        id:
          type: string
          format: ObjectId
        user_id:
          type: string
          format: ObjectId
        quiz_id:
          type: string
          format: ObjectId
        module_id:
          type: string
          format: ObjectId
        date_completed:
          type: string
          format: date
        user_answers:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionUserAnswers'
        grade:
          type: integer
          format: int32
      example:
        quiz_id: quizId
        user_answers:
          - question_id: questionId
            user_answer: foo
          - question_id: questionId
            user_answer: bar
        date_completed: '2019-09-13 15:47:18.171396'
        grade: 6
        module_id: moduleId
        user_id: userId
        id: someId
    PostSubmission:
      description: Content received when a user gets a submission
      type: object
      properties:
        user_id:
          type: string
          format: ObjectId
        quiz_id:
          type: string
          format: ObjectId
        module_id:
          type: string
          format: ObjectId
        user_answers:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionUserAnswers'
      required:
        - user_id
        - quiz_id
        - module_id
        - user_answers
      example:
        module_id: moduleId
        user_id: userId
        quiz_id: quizId
        user_answers:
          - question_id: questionId
            user_answer: foo
          - question_id: questionId
            user_answer: bar
    SubmissionUserAnswers:
      description: Object containing user answer to a question with the questions answer Id
      type: object
      properties:
        question_id:
          type: string
          format: ObjectId
        user_answer:
          type: string
      required:
        - questionId
        - user_answer
    BulkCreateQuiz:
      description: Quiz object
      type: object
      properties:
        quiz_name:
          type: string
        details:
          type: string
        questions:
          type: array
          items:
            type: object
            properties:
              question_text:
                type: string
              url:
                type: string
                format: uri
              answer_bank:
                type: array
                items:
                  type: string
              correct_index:
                type: integer
                format: int32
...
    PostQuestion:
      description: Question object
      type: object
      properties:
        question_text:
          type: string
        word:
          type: string
          format: ObjectId
      example:
        question_text: Hi this is a test
        word: someId
    GetQuestion:
      allOf:
        - $ref: '#/components/schemas/PostQuestion'
      type: object
      properties:
        id:
          type: string
          format: ObjectId
      example:
        id: 'someId'
    PostQuiz:
      description: Content received when a user gets a submission
      type: object
      properties:
        quiz_name:
          type: string
        details:
          type: string
      example:
        quiz_name: Test quiz
        details: This is a test quiz
    GetQuiz:
      allOf:
        - $ref: '#/components/schemas/PostQuiz'
      type: object
      properties:
        id:
          type: string
          format: ObjectId
      example:
        id: 'someId'